requirements.txt
# pip install -r requirements.txt
python-dotenv
pyyaml
google-api-python-client==2.88.0
tweepy==4.14.0
requests==2.31.0
instagrapi==1.18.4
pyquery

